refactor(core): merge types.ts into create-app.ts
Types follow functions - all app types now live with createApp.
Eliminates types.ts which had definitions without usage.

Verified: precommit-verify ✓ (ac3f39670a59f6b8)
