+++ packages/core/src/app/index.ts
@@ -1,9 +1,9 @@
-export { createApp } from './create-app';
-export type {
-  App,
-  HttpApp,
-  CommandApp,
-  CreateAppOptions,
-  ReadyOptions,
-  ReadyResult,
-} from './types';
+export {
+  createApp,
+  type App,
+  type HttpApp,
+  type CommandApp,
+  type CreateAppOptions,
+  type ReadyOptions,
+  type ReadyResult,
+} from './create-app';
